# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bhojnalaya API",
        "description": "REST API documentation for tables, menu items, orders, and order items.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/menu-items": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "List menu items",
                "operationId": "4ebff4229af20351ff03e10a24a2fee5",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Starters"
                        }
                    },
                    {
                        "name": "is_available",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu items fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MenuItem"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Create a menu item",
                "operationId": "66c0502470d5f2b36a95aa1310af7ab8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MenuItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuItem"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/menu-items/{menu_item}": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Get a single menu item",
                "operationId": "9aa5713e6fb2bad003ae346861e62e32",
                "parameters": [
                    {
                        "name": "menu_item",
                        "in": "path",
                        "description": "Menu item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu item fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuItem"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Update a menu item",
                "operationId": "12af2e566bd8cd44e678b9925b4aeb26",
                "parameters": [
                    {
                        "name": "menu_item",
                        "in": "path",
                        "description": "Menu item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MenuItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuItem"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Delete a menu item",
                "operationId": "f4170c53c8d994a745434a3c0d23679f",
                "parameters": [
                    {
                        "name": "menu_item",
                        "in": "path",
                        "description": "Menu item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Menu item deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MenuItem": {
                "required": [
                    "id",
                    "name",
                    "category",
                    "item_type",
                    "price",
                    "is_available"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Paneer Tikka"
                    },
                    "category": {
                        "type": "string",
                        "example": "Starters"
                    },
                    "short_code": {
                        "type": "string",
                        "example": "PT01",
                        "nullable": true
                    },
                    "item_type": {
                        "type": "string",
                        "example": "Veg"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 220
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:05:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-15T10:05:00.000000Z"
                    }
                },
                "type": "object"
            },
            "MenuItemRequest": {
                "required": [
                    "name",
                    "category",
                    "item_type",
                    "price"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Paneer Tikka"
                    },
                    "category": {
                        "type": "string",
                        "example": "Starters"
                    },
                    "short_code": {
                        "type": "string",
                        "example": "PT01"
                    },
                    "item_type": {
                        "type": "string",
                        "example": "Veg"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 220
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The name field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
...
}